Refactor server creation/routing to worker-specific file
