+++ frameworks/JavaScript/nodejs/app.js
@@ -1,0 +1,16 @@
+var cluster = require('cluster');
+var numCPUs = require('os').cpus().length;
+
+if (cluster.isMaster) {
+  // Fork workers.
+  for (var i = 0; i < numCPUs; i++) {
+    cluster.fork();
+  }
+
+  cluster.on('exit', function (worker, code, signal) {
+    console.log('worker ' + worker.pid + ' died');
+  });
+} else {
+  // Task for forked worker
+  require('./create-server');
+}
